finish integrating sleepy detector (fatigue or active) to the CameraActivity.kt

+++ app/src/main/java/id/ac/unri/driverfit/domain/model/Classification.kt
@@ -2,5 +2,5 @@ package id.ac.unri.driverfit.domain.model
 
 data class Classification(
     val label: String,
-    val scores: Int,
+    val scores: Float,
 )
